requirements.txt
Flask==0.10.1
Flask-RESTful==0.3.5
Flask-SQLAlchemy==2.1
Flask-WTF==0.12
IP2Location==7.0.4
Jinja2==2.8
MarkupSafe==0.23
SQLAlchemy==1.0.12
Tempita==0.5.3dev
WTForms==2.1
Werkzeug==0.11.4
aniso8601==1.1.0
argparse==1.2.1
decorator==4.0.9
distribute==0.6.24
itsdangerous==0.24
numpy==1.11.0rc1
pandas==0.18.0
pbr==1.8.1
pyephem==3.7.6.0
pyswisseph==2.00.00-2
python-dateutil==2.5.0
pytz==2016.1
six==1.10.0
sqlalchemy-migrate==0.10.0
sqlparse==0.1.19
wsgiref==0.1.2
